+++ demo_projects/project_C/add.c
@@ -1,0 +1,10 @@
+#include "project_C/add.h"
+
+#ifndef YYY_defined
+compile error
+#endif
+
+int add(int a, int b)
+{
+    return a + b;
+}